+++ tests/hpc/test-exports.R
@@ -25,14 +25,17 @@ tar_test("runtime settings get exported to workers", {
   skip_if_not_installed("clustermq")
   tar_script({
     options(clustermq.scheduler = "multiprocess")
+    Sys.setenv(TAR_WARN = "success")
     list(
       tar_target(x, targets::tar_store()),
-      tar_target(y, targets::tar_function())
+      tar_target(y, targets::tar_function()),
+      tar_target(z, Sys.getenv("TAR_WARN"))
     )
   })
   tar_make_clustermq(store = "local_store")
   expect_equal(tar_read(x, store = "local_store"), "local_store")
   expect_equal(tar_read(y, store = "local_store"), "tar_make_clustermq")
+  expect_equal(tar_read(z, store = "local_store"), "success")
   expect_null(tar_function())
   expect_false(tar_store() == "local_store")
 })
@@ -41,14 +44,17 @@ tar_test("same with the future backend", {
   skip_if_not_installed("future")
   tar_script({
     future::plan(future::multisession)
+    Sys.setenv(TAR_WARN = "success")
     list(
       tar_target(x, targets::tar_store()),
-      tar_target(y, targets::tar_function())
+      tar_target(y, targets::tar_function()),
+      tar_target(z, Sys.getenv("TAR_WARN"))
     )
   })
   tar_make_future(store = "local_store")
   expect_equal(tar_read(x, store = "local_store"), "local_store")
   expect_equal(tar_read(y, store = "local_store"), "tar_make_future")
+  expect_equal(tar_read(z, store = "local_store"), "success")
   expect_null(tar_function())
   expect_false(tar_store() == "local_store")
 })