+++ NEWS.md
@@ -4,6 +4,7 @@
 * Set the `store` field of `tar_runtime` prior to sourcing `_targets.R` so `tar_store()` works in target scripts.
 * Add a new function `tar_function()` to return the `targets` function currently running (from `_targets.R` or a target).
 * Add a new function `tar_active()` to tell whether the pipeline is currently running. Detects if it is called from `tar_make()` or similar function.
+* Explicitly export all the environment variables from `tar_envvars()` to targets run on parallel workers.
 
 # targets 0.9.1